tore out global manager as i found out about object_id on Object class. makes a lit more sense to use that for comparisons and such
